pyproject.toml
[project]
name = "oktal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.1",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "dacite>=1.9.2",
    "debugpy>=1.8.14",
    "duckduckgo-search>=8.0.2",
    "fastapi[standard]>=0.115.12",
    "greenlet>=3.2.2",
    "icecream>=2.1.4",
    "langchain>=0.3.25",
    "langchain-core>=0.3.63",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.19",
    "langgraph>=0.4.8",
    "langgraph-checkpoint>=2.0.26",
    "langgraph-checkpoint-sqlite>=2.0.10",
    "loguru>=0.7.3",
    "openai>=1.82.1",
    "passlib>=1.7.4",
    "pgvector>=0.4.1",
    "psycopg>=3.2.9",
    "pydantic-settings>=2.9.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.41",
    "structlog>=25.3.0",
    "tavily-python>=0.7.3",
    "tiktoken>=0.9.0",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.14",
    "isort>=6.0.1",
    "lxml-stubs>=0.5.1",
    "mypy>=1.16.0",
    "pandas-stubs>=2.2.3.250527",
    "ruff>=0.11.12",
    "ruff-lsp>=0.0.62",
    "types-beautifulsoup4>=4.12.0.20250516",
    "types-colorama>=0.4.15.20240311",
    "types-defusedxml>=0.7.0.20250516",
    "types-greenlet>=3.2.0.20250417",
    "types-jsonschema>=4.24.0.20250528",
    "types-passlib>=1.7.7.20250602",
    "types-psutil>=7.0.0.20250601",
    "types-psycopg2>=2.9.21.20250516",
    "types-pygments>=2.19.0.20250516",
    "types-python-jose>=3.5.0.20250531",
    "types-pyyaml>=6.0.12.20250516",
    "types-regex>=2024.11.6.20250403",
    "types-requests>=2.32.0.20250602",
    "types-tqdm>=4.67.0.20250516",
    "types-ujson>=5.10.0.20250326"
]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
strict_optional = true
disallow_untyped_defs = true
warn_unused_ignores = true

[tool.ruff]
extend-select = ["I"]